RCNN improvements (#901)
* bug fix for RCNN
1. Encode bbox using float instead of int.
2. generate anchor using detectron's strategy.
3. additional comment to clarify code.

* add multi thread executor. improve speed on p3.24x instance

* rm unused arg

* missing docstring

* add groupnorm back. groupnorm will be added to mxnet in 1.6

* bug fix

rename model

rename model

fix

fix

* refactor rcnn metrics

* add docs string

* expose executor threads. this has significant effect on performance. the setting also needs to be tune depends on the machine

* typo

+++ gluoncv/data/transforms/bbox.py
@@ -126,7 +126,7 @@ def resize(bbox, in_size, out_size):
     if not len(out_size) == 2:
         raise ValueError("out_size requires length 2 tuple, given {}".format(len(out_size)))
 
-    bbox = bbox.copy()
+    bbox = bbox.copy().astype(float)
     x_scale = out_size[0] / in_size[0]
     y_scale = out_size[1] / in_size[1]
     bbox[:, 1] = y_scale * bbox[:, 1]
